# Made bloom_filter type of index supporting LowCardinality and Nullable