--- a/dbms/src/Functions/FunctionHelpers.cpp
+++ b/dbms/src/Functions/FunctionHelpers.cpp
@@ -4,8 +4,10 @@
 #include <Columns/ColumnString.h>
 #include <Columns/ColumnFixedString.h>
 #include <Columns/ColumnNullable.h>
+#include <Columns/ColumnLowCardinality.h>
 #include <Common/assert_cast.h>
 #include <DataTypes/DataTypeNullable.h>
+#include <DataTypes/DataTypeLowCardinality.h>
 #include <IO/WriteHelpers.h>
 
 
@@ -74,6 +76,12 @@ static Block createBlockWithNestedColumnsImpl(const Block & block, const std::un
             {
                 const auto & nested_col = checkAndGetColumn<ColumnNullable>(const_column->getDataColumn())->getNestedColumnPtr();
                 res.insert({ ColumnConst::create(nested_col, col.column->size()), nested_type, col.name});
+            }
+            else if (auto * low_cardinality = checkAndGetColumn<ColumnLowCardinality>(*col.column))
+            {
+                const DataTypePtr & low_cardinality_type = static_cast<const DataTypeLowCardinality &>(*col.type).getDictionaryType();
+                const auto & low_cardinality_col = low_cardinality->convertToFullColumnIfLowCardinality();
+                res.insert({low_cardinality_col, low_cardinality_type, col.name});
             }
             else
                 throw Exception("Illegal column for DataTypeNullable", ErrorCodes::ILLEGAL_COLUMN);