--- a/dbms/src/Storages/MergeTree/MergeTreeIndexBloomFilter.cpp
+++ b/dbms/src/Storages/MergeTree/MergeTreeIndexBloomFilter.cpp
@@ -12,6 +12,7 @@
 #include <Storages/MergeTree/MergeTreeIndexConditionBloomFilter.h>
 #include <Parsers/queryToString.h>
 #include <Columns/ColumnConst.h>
+#include <Columns/ColumnLowCardinality.h>
 #include <Interpreters/BloomFilterHash.h>
 
 
@@ -74,13 +75,8 @@ static void assertIndexColumnsType(const Block & header)
 
     for (size_t index = 0; index < columns_data_types.size(); ++index)
     {
-        WhichDataType which(columns_data_types[index]);
-
-        if (which.isArray())
-        {
-            const DataTypeArray * array_type = typeid_cast<const DataTypeArray *>(columns_data_types[index].get());
-            which = WhichDataType(array_type->getNestedType());
-        }
+        const IDataType * actual_type = getPrimitiveType(columns_data_types[index]).get();
+        WhichDataType which(actual_type);
 
         if (!which.isUInt() && !which.isInt() && !which.isString() && !which.isFixedString() && !which.isFloat() &&
             !which.isDateOrDateTime() && !which.isEnum())