I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Category:
* New Feature

Short description:
Since current bloom_filter doesn't support LowCardinality and Nullable, so I made bloom_filter type of index supporting LowCardinality and Nullable.

### To use bloom_filter() for LowCardinality and Nullable, which I developed

1. Create a table with bloom_filter for LowCardinality and Nullable.

```SQL
SET allow_experimental_data_skipping_indices = 1;

CREATE TABLE default.test_bloomfilter
(
    `key` LowCardinality(Nullable(String)),
    `key_no_idx` LowCardinality(Nullable(String)),
    `id` String,
    `dt` DateTime,
    INDEX key_bf key TYPE bloom_filter(0.01) GRANULARITY 8192
)
ENGINE = MergeTree()
PARTITION BY toYYYYMMDD(dt)
ORDER BY id;
```

2. Insert data to the table.
```SQL
INSERT INTO test_bloomfilter VALUES ('k1', 'k1', 'id1', '2019-01-01 00:00:00'); 
INSERT INTO test_bloomfilter VALUES ('k2', 'k2', 'id1', '2019-01-01 00:00:00'); 
INSERT INTO test_bloomfilter VALUES ('k3', 'k3', 'id1', '2019-01-01 00:00:00'); 
INSERT INTO test_bloomfilter VALUES ('k4', 'k4', 'id1', '2019-01-01 00:00:00'); 
INSERT INTO test_bloomfilter VALUES ('k5', 'k5', 'id1', '2019-01-01 00:00:00'); 
INSERT INTO test_bloomfilter VALUES (NULL, NULL, 'id1', '2019-01-01 00:00:00'); 
```

3. Select data with bloom_filter.
```SQL
SELECT * from test_bloomfilter where key = 'k1';
SELECT * from test_bloomfilter where isNull(key);
```

### Performance Test
I did `Group by` performance test.
As you know, the `Group by` operation on LowCardinality outperforms the non LowCardinality.
It is available for a column with LowCardinality and bloom_filter to  use `fast filtering by bloom_filter` and `fast grouping by LowCardinality`

```SQL
CREATE TABLE default.perf_bloomfilter                       
(                                                           
    `d1` LowCardinality(Nullable(String)),                  
    `d2` String,                                            
    `id` String,                                            
    `dt` DateTime,                                          
    INDEX idx_bf d1 TYPE bloom_filter(0.01) GRANULARITY 8192
)                                                           
ENGINE = MergeTree()                                        
PARTITION BY toYYYYMMDD(dt)                                 
ORDER BY id                                                 
```

```SQL
INSERT INTO perf_bloomfilter SELECT
    r,                             
    r,                             
    r AS id,                       
    toDateTime(p) AS dt            
FROM                               
(                                  
    SELECT                         
        number % 10 AS r,          
        number % 1000000 AS p      
    FROM system.numbers            
    LIMIT 10000000                 
)                                  
```

**Without LowCardinality**
```SQL
SELECT               
    d2,              
    count()          
FROM perf_bloomfilter
GROUP BY d2          
```
100 rows in set. Elapsed: `0.032 sec`. Processed 10.00 million rows, 109.00 MB (316.08 million rows/s., 3.45 GB/s.)

**With LowCardinality**
```SQL
SELECT               
    d1,              
    count()          
FROM perf_bloomfilter
GROUP BY d1          
```
100 rows in set. Elapsed: `0.019 sec`. Processed 10.00 million rows, 10.02 MB (516.15 million rows/s., 517.05 MB/s.)