@achimbab 2019-10-23
@zhang2014 
Hello~ 
I appreciate your previous code review #6984.
And nice to met you again. 
I fixed codes to pass tests, expecting your review. :)